implement admin layout, bulletin form component

+++ WebApp/ClientApp/src/admin/components/bulletin/editor/editor.component.ts
@@ -3,6 +3,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 
 import { BulletinEditDto } from '../../../../interfaces/bulletin.interfaces';
 import { AdminBulletinService } from '../../../services/bulletin.service';
+import { faEdit, faTrash } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
   selector: 'app-admin-bulletin-editor',
@@ -10,6 +11,9 @@ import { AdminBulletinService } from '../../../services/bulletin.service';
   styleUrls: ['./editor.component.css']
 })
 export class AdminBulletinEditorComponent implements OnInit {
+  faEdit = faEdit;
+  faTrash = faTrash;
+
   public edit: boolean;
   public bulletinId: number;
   public bulletin: BulletinEditDto;
@@ -31,6 +35,7 @@ export class AdminBulletinEditorComponent implements OnInit {
   public editBulletin() {
     this.edit = true;
     this.router.navigate(['/admin/bulletin', this.bulletinId], {
+      replaceUrl: true,
       queryParams: { edit: true }
     });
   }
@@ -38,13 +43,15 @@ export class AdminBulletinEditorComponent implements OnInit {
   public updateBulletin(bulletin: BulletinEditDto) {
     this.service.updateSingle(bulletin).subscribe(() => {
       this.edit = false;
-      this.router.navigate(['/admin/bulletin', this.bulletinId]);
+      this.router.navigate(['/admin/bulletin', this.bulletinId], { replaceUrl: true });
     }, error => console.error(error));
   }
 
   public deleteBulletin() {
-    this.service.deleteSingle(this.bulletinId).subscribe(() => {
-      this.router.navigate(['/admin/bulletin']);
-    }, error => console.error(error));
+    if (confirm('Are you sure to delete bulletin #' + this.bulletinId + '?')) {
+      this.service.deleteSingle(this.bulletinId).subscribe(() => {
+        this.router.navigate(['/admin/bulletin']);
+      }, error => console.error(error));
+    }
   }
 }
